Refactor backprop traversal & other improvements
The main change is a complete refactoring of network travesal during
backprop. It used a complex recursive method before but now handles it's
own custom stack and iterates. Probably use similar traversal in
Network#evaluate also, after benching.

We're also now using Parallels "finish" option which acts like a reduce
step and merges our batch results in the main thread as and when the
batch completes, rather than doing it after processing all batches. This
is also much more memory efficent as it doesn't need to enumerate all
the results from the batches in memory.

Bugfixes:

 - Backprop for ProductNeurons was using the intermediate value of the
   final timestep rather than the correct timestep. It's now returned in
   Network#state along with the neuron values.
 - There was an index bug in the parallel section's batch handling.
 - Probably either fixed or broke some other stuff during the refactor,
   too.